device/requirements.txt
python-dotenv==1.0.1
paho-mqtt==2.1.0
pydantic==2.9.2